+++ api/v1/bootstrap.php
@@ -1,0 +1,5 @@
+<?php
+if (!defined('APP_API_VERSION')) {
+    define('APP_API_VERSION', 'v1');
+    define('APP_UI_VERSION', '1.1.1');
+}