+++ api/v2/bootstrap.php
@@ -1,0 +1,5 @@
+<?php
+if (!defined('APP_API_VERSION')) {
+    define('APP_API_VERSION', 'v2');
+    define('APP_UI_VERSION', '2.0.0');
+}